clarify exactly which files we need to format after a lint failure in the pre-commit hook
