+++ build-support/BUILD
@@ -1,0 +1,4 @@
+files(
+    name='common-script',
+    sources=['common.sh'],
+)
